Organize CLI modules by ownership
