+++ wmo/cli/config/__init__.py
@@ -1,0 +1,1 @@
+"""Project-local configuration command modules."""
