+++ wmo/cli/judge/__init__.py
@@ -1,0 +1,1 @@
+"""Manual judge setup, calibration, and review command modules."""
