+++ wmo/cli/providers/__init__.py
@@ -1,0 +1,1 @@
+"""Provider discovery and model-catalog setup modules shared by CLI commands."""
