+++ wmo/cli/shared/__init__.py
@@ -1,0 +1,1 @@
+"""Terminal, Typer, and progress primitives shared by CLI command packages."""
