+++ wmo/cli/shared/consent.py
