+++ wmo/cli/shared/options_test.py
