+++ wmo/conftest.py
@@ -26,7 +26,7 @@ _SETTLE_SECONDS = 0.3
 @pytest.fixture
 def interactive_stdin(monkeypatch: pytest.MonkeyPatch) -> None:
     """Present a terminal stdin for a test that must reach an interactive prompt."""
-    from wmo.cli import consent
+    from wmo.cli.shared import consent
 
     monkeypatch.setattr(consent, "_stdin_is_terminal", lambda: True)
